requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# LLM & RAG
langchain==0.3.1
langchain-openai==0.2.1
langchain-anthropic==0.2.1
langchain-community==0.3.1
langchain-chroma==0.1.4
langgraph==0.2.28

# Vector Store
chromadb==0.5.15

# GitHub Ingestion
PyGithub==2.4.0
gitpython==3.1.43

# Embeddings & ML
openai==1.51.0
anthropic==0.34.2
tiktoken==0.7.0
numpy==1.26.4

# Utilities
httpx==0.27.2
rich==13.8.1
tenacity==9.0.0
